ok god bless peers, unsigned long changed to unsigned int and now everythings perfectgit add .! lets work on return now

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mratke <mratke@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/17 12:14:19 by mratke            #+#    #+#             */
-/*   Updated: 2024/10/18 18:18:39 by mratke           ###   ########.fr       */
+/*   Updated: 2024/10/19 18:38:20 by mratke           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,5 +27,5 @@ int	ft_printf_int(int n);
 int	ft_printf_uint(unsigned int n);
 int	ft_printf_percent(void);
 int	ft_printf_ptr(void *ptr);
-int	ft_printf_hexadecimal(long int n, int type);
+int	ft_printf_hexadecimal(unsigned int n, int type);
 #endif
